Split Swagger2Translator in two (restlet/apispark#1166)
