+++ modules/org.restlet.test/src/org/restlet/test/ext/apispark/conversion/swagger/v2_0/Swagger2TranslatorTestCase.java
@@ -32,7 +32,8 @@ import java.util.Map;
 
 import org.restlet.data.ChallengeScheme;
 import org.restlet.data.MediaType;
-import org.restlet.ext.apispark.internal.conversion.swagger.v2_0.Swagger2Translator;
+import org.restlet.ext.apispark.internal.conversion.swagger.v2_0.Swagger2Reader;
+import org.restlet.ext.apispark.internal.conversion.swagger.v2_0.Swagger2Writer;
 import org.restlet.ext.apispark.internal.model.Contract;
 import org.restlet.ext.apispark.internal.model.Definition;
 import org.restlet.ext.apispark.internal.model.Endpoint;
@@ -225,7 +226,7 @@ public class Swagger2TranslatorTestCase extends Swagger2TestCase {
         representation2.setName("nameRepresentation2");
 
         // When
-        Swagger swagger = Swagger2Translator.getSwagger(definition);
+        Swagger swagger = Swagger2Reader.getSwagger(definition);
 
         // Then
         assertEquals("2.0", swagger.getSwagger());
@@ -360,7 +361,7 @@ public class Swagger2TranslatorTestCase extends Swagger2TestCase {
                         .getFile(), MediaType.APPLICATION_JSON),
                 Definition.class).getObject();
 
-        Swagger translatedSwagger = Swagger2Translator
+        Swagger translatedSwagger = Swagger2Reader
                 .getSwagger(savedDefinition);
 
         URL refImpl = getClass().getResource("refImpl.swagger");
@@ -373,7 +374,7 @@ public class Swagger2TranslatorTestCase extends Swagger2TestCase {
         URL refImpl = getClass().getResource("refImpl.swagger");
         Swagger savedSwagger = SwaggerLoader.readJson(refImpl.getFile());
         
-        Definition translatedDefinition = Swagger2Translator.translate(savedSwagger);
+        Definition translatedDefinition = Swagger2Writer.translate(savedSwagger);
         
         Definition savedDefinition = new JacksonRepresentation<>(
                 new FileRepresentation(getClass().getResource("refImpl.rwadef")
